pyproject.toml
[tool.poetry]
name = "smart"
version = "0.1.0"
description = ""
authors = ["xjoerootx <joeroot1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
celery = "^5.4.0"
minio = "^7.2.15"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.11"
alembic = "^1.14.1"
docker = "^7.1.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
asyncssh = "^2.19.0"
aiofiles = "^24.1.0"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
